Created Player base class
- Updated Networking scene organization
- Converted Player to base class
- Created PlayerDesktop class for desktop users
- Added Player container to server
