+++ Assets/Project/Prefabs/Player/ServerSidePlayer.prefab
@@ -42,15 +42,15 @@ MonoBehaviour:
   m_GameObject: {fileID: 2329821471381548265}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9efb3adc5a4d54418aca2d127d6eb903, type: 3}
+  m_Script: {fileID: 11500000, guid: 3a392676881c6d044ab6ebe860bf2721, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   id: 0
   username: 
   controller: {fileID: 1037102014086549081}
-  gravity_per_tick: -9.81
+  gravity_per_tick: -19.62
   moveSpeed_per_tick: 5
-  jumpSpeed_per_tick: 5
+  jumpSpeed_per_tick: 9
 --- !u!143 &1037102014086549081
 CharacterController:
   m_ObjectHideFlags: 0
