+++ Assets/Project/Scenes/NetworkingDemo-Server.unity
@@ -572,6 +572,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   playerPrefab: {fileID: 2329821471381548265, guid: 018a3bf8abddb4eb9a7a8c526d0050d0,
     type: 3}
+  playersContainer: {fileID: 1928696140}
 --- !u!4 &809287217
 Transform:
   m_ObjectHideFlags: 0
