+++ Assets/Project/Scripts/NetworkingScripts/Server/PlayerDesktop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a392676881c6d044ab6ebe860bf2721
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
